requirements.txt
fastapi>=0.100.0
titiler.core==0.21.1
titiler.xarray==0.21.1
titiler.extensions==0.21.1
titiler.mosaic==0.21.1
titiler.application==0.21.1
uvicorn
gunicorn
quantized-mesh-encoder>=0.4.3 
numpy==1.26.4
pymartini>=0.4.4
pydelatin>=0.2.7
morecantile>=3.2.5